// src/app/components/models/all.models.ts
export interface Evaluation {
  id?: number;
  name: string;           // name of the evaluation type e.g., "Midterm Math"
  description?: string;   // optional longer description
  /*maxScore?: number;      // maximum attainable score
  passingScore?: number;  // score required to pass*/
  active?: boolean;       // whether this evaluation template is active
  //remarks?: string;
}

export interface EvaluationDTO {
  name: string;
  description?: string;
  active?: boolean;
}